Can now specify array for preferred operator
You can now either specify a string, to use a preferred operator, or a
list. If you specify a list, it loops through it in reverse order and
picks the last one (closest to the front) that is available. No real
point in optimizing and finding the first available and breaking out
of the loop b/c I don't expect many people to provide hundreds of
preferred operators.